Added src port binding to the client.
I did this so that the listener which knows about the magic port number can
use it to identify if the external port of a node is being randomized or is
different than them magic src port. This is useful for determinig which node
should be the p2p listener and which should be the p2p initiator.